+++ samples/SimpleWebApp/Controllers/ValuesController.cs
@@ -25,7 +25,7 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using Microsoft.AspNetCore.Mvc;
 
-namespace WebApplication.Controllers
+namespace SimpleWebApp.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
