+++ samples/SimpleWebApp/Views/Home/Index.cshtml
