--- samples/WebApplication/appsettings.Development.json
@@ -1,9 +1,0 @@
-{
-    "Logging": {
-        "LogLevel": {
-            "Default": "Debug",
-            "System": "Information",
-            "Microsoft": "Information"
-        }
-    }
-}
