+++ src/AspNetCore.AutoHealthCheck.sln
@@ -11,19 +11,19 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{DFA2E39C-2C5
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{270C7C3F-98AE-43D8-BB6B-18FAE0420F89}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AspNetCore.AutoHealthCheck.Tests", "AspNetCore.AutoHealthCheck.Tests\AspNetCore.AutoHealthCheck.Tests.csproj", "{32641493-0310-4380-B7CF-47F55D74F0A9}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebApplication", "..\samples\WebApplication\WebApplication.csproj", "{419C8FDE-17BA-42DF-AD2A-DF19323E706D}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AspNetCoreUsingDiagnostic", "..\samples\AspNetCoreUsingDiagnostic\AspNetCoreUsingDiagnostic.csproj", "{9E500F6B-4F0A-440E-8392-BB316C8D4048}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AspNetCore.AutoHealthCheck.Diagnostics", "AspNetCore.AutoHealthCheck.Diagnostics\AspNetCore.AutoHealthCheck.Diagnostics.csproj", "{A5D81A52-319E-489D-929E-5F285495938C}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AspNetCore.AutoHealthCheck.Abstractions", "AspNetCore.AutoHealthCheck.Abstractions\AspNetCore.AutoHealthCheck.Abstractions.csproj", "{7A4F34EB-A8AF-4A26-844F-DA2C52121622}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AspNetCore.AutoHealthCheck.Diagnostics.Tests", "AspNetCore.AutoHealthCheck.Diagnostics.Tests\AspNetCore.AutoHealthCheck.Diagnostics.Tests.csproj", "{A15EA5B1-8E56-437A-8FC8-93ECBB9E78F9}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AspNetCoreUsingAutoHealthCheck", "..\samples\AspNetCoreUsingAutoHealthCheck\AspNetCoreUsingAutoHealthCheck.csproj", "{79CCC024-FF8D-4F15-BB25-7963A306D560}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleWebApp", "..\samples\SimpleWebApp\SimpleWebApp.csproj", "{39F6B861-263B-4416-9818-976EF66AEBCF}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AspNetCore.AutoHealthCheck.Diagnostics.Tests", "..\tests\AspNetCore.AutoHealthCheck.Diagnostics.Tests\AspNetCore.AutoHealthCheck.Diagnostics.Tests.csproj", "{D00EFD27-A3C3-401F-A1A6-FAFB3D241BBB}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AspNetCore.AutoHealthCheck.Tests", "..\tests\AspNetCore.AutoHealthCheck.Tests\AspNetCore.AutoHealthCheck.Tests.csproj", "{1A8DB673-6DF8-410D-81C7-CF837BB595AC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -35,14 +35,6 @@ Global
 		{B2553537-A52E-49A9-A36F-C42F5D441EEA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B2553537-A52E-49A9-A36F-C42F5D441EEA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B2553537-A52E-49A9-A36F-C42F5D441EEA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{32641493-0310-4380-B7CF-47F55D74F0A9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{32641493-0310-4380-B7CF-47F55D74F0A9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{32641493-0310-4380-B7CF-47F55D74F0A9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{32641493-0310-4380-B7CF-47F55D74F0A9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{419C8FDE-17BA-42DF-AD2A-DF19323E706D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{419C8FDE-17BA-42DF-AD2A-DF19323E706D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{419C8FDE-17BA-42DF-AD2A-DF19323E706D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{419C8FDE-17BA-42DF-AD2A-DF19323E706D}.Release|Any CPU.Build.0 = Release|Any CPU
 		{9E500F6B-4F0A-440E-8392-BB316C8D4048}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{9E500F6B-4F0A-440E-8392-BB316C8D4048}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9E500F6B-4F0A-440E-8392-BB316C8D4048}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -55,27 +47,35 @@ Global
 		{7A4F34EB-A8AF-4A26-844F-DA2C52121622}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7A4F34EB-A8AF-4A26-844F-DA2C52121622}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7A4F34EB-A8AF-4A26-844F-DA2C52121622}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A15EA5B1-8E56-437A-8FC8-93ECBB9E78F9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A15EA5B1-8E56-437A-8FC8-93ECBB9E78F9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A15EA5B1-8E56-437A-8FC8-93ECBB9E78F9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A15EA5B1-8E56-437A-8FC8-93ECBB9E78F9}.Release|Any CPU.Build.0 = Release|Any CPU
 		{79CCC024-FF8D-4F15-BB25-7963A306D560}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{79CCC024-FF8D-4F15-BB25-7963A306D560}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{79CCC024-FF8D-4F15-BB25-7963A306D560}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{79CCC024-FF8D-4F15-BB25-7963A306D560}.Release|Any CPU.Build.0 = Release|Any CPU
+		{39F6B861-263B-4416-9818-976EF66AEBCF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{39F6B861-263B-4416-9818-976EF66AEBCF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{39F6B861-263B-4416-9818-976EF66AEBCF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{39F6B861-263B-4416-9818-976EF66AEBCF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D00EFD27-A3C3-401F-A1A6-FAFB3D241BBB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D00EFD27-A3C3-401F-A1A6-FAFB3D241BBB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D00EFD27-A3C3-401F-A1A6-FAFB3D241BBB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D00EFD27-A3C3-401F-A1A6-FAFB3D241BBB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1A8DB673-6DF8-410D-81C7-CF837BB595AC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1A8DB673-6DF8-410D-81C7-CF837BB595AC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1A8DB673-6DF8-410D-81C7-CF837BB595AC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1A8DB673-6DF8-410D-81C7-CF837BB595AC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{B2553537-A52E-49A9-A36F-C42F5D441EEA} = {DFA2E39C-2C56-42DE-8833-C36C392A07A7}
-		{32641493-0310-4380-B7CF-47F55D74F0A9} = {270C7C3F-98AE-43D8-BB6B-18FAE0420F89}
-		{419C8FDE-17BA-42DF-AD2A-DF19323E706D} = {78D3E477-EE1C-4D6D-ADE9-366A4D7C6560}
 		{9E500F6B-4F0A-440E-8392-BB316C8D4048} = {78D3E477-EE1C-4D6D-ADE9-366A4D7C6560}
 		{A5D81A52-319E-489D-929E-5F285495938C} = {DFA2E39C-2C56-42DE-8833-C36C392A07A7}
 		{7A4F34EB-A8AF-4A26-844F-DA2C52121622} = {DFA2E39C-2C56-42DE-8833-C36C392A07A7}
-		{A15EA5B1-8E56-437A-8FC8-93ECBB9E78F9} = {270C7C3F-98AE-43D8-BB6B-18FAE0420F89}
 		{79CCC024-FF8D-4F15-BB25-7963A306D560} = {78D3E477-EE1C-4D6D-ADE9-366A4D7C6560}
+		{39F6B861-263B-4416-9818-976EF66AEBCF} = {78D3E477-EE1C-4D6D-ADE9-366A4D7C6560}
+		{D00EFD27-A3C3-401F-A1A6-FAFB3D241BBB} = {270C7C3F-98AE-43D8-BB6B-18FAE0420F89}
+		{1A8DB673-6DF8-410D-81C7-CF837BB595AC} = {270C7C3F-98AE-43D8-BB6B-18FAE0420F89}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {64B4BA9D-729D-4F8A-BBCF-F3E3FDA7E599}
