+++ tests/AspNetCore.AutoHealthCheck.Diagnostics.Tests/AspNetCore.AutoHealthCheck.Diagnostics.Tests.csproj
@@ -14,8 +14,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\AspNetCore.AutoHealthCheck.Abstractions\AspNetCore.AutoHealthCheck.Abstractions.csproj" />
-    <ProjectReference Include="..\AspNetCore.AutoHealthCheck.Diagnostics\AspNetCore.AutoHealthCheck.Diagnostics.csproj" />
+    <ProjectReference Include="..\..\src\AspNetCore.AutoHealthCheck.Abstractions\AspNetCore.AutoHealthCheck.Abstractions.csproj" />
+    <ProjectReference Include="..\..\src\AspNetCore.AutoHealthCheck.Diagnostics\AspNetCore.AutoHealthCheck.Diagnostics.csproj" />
   </ItemGroup>
 
 </Project>
