+++ tests/AspNetCore.AutoHealthCheck.Diagnostics.Tests/AspNetCoreDiagnosticHealthCheckTests.cs
