+++ tests/AspNetCore.AutoHealthCheck.Tests/AspNetCore.AutoHealthCheck.Tests.csproj
@@ -17,7 +17,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\AspNetCore.AutoHealthCheck\AspNetCore.AutoHealthCheck.csproj" />
+    <ProjectReference Include="..\..\src\AspNetCore.AutoHealthCheck.Abstractions\AspNetCore.AutoHealthCheck.Abstractions.csproj" />
+    <ProjectReference Include="..\..\src\AspNetCore.AutoHealthCheck\AspNetCore.AutoHealthCheck.csproj" />
   </ItemGroup>
 
 </Project>
