+++ tests/AspNetCore.AutoHealthCheck.Tests/Controllers/FakeController.cs
