+++ tests/AspNetCore.AutoHealthCheck.Tests/Infrastructure/CustomProbes/ProbeResultTests.cs
