+++ tests/AspNetCore.AutoHealthCheck.Tests/Infrastructure/EndpointBuilderTests.cs
