+++ tests/AspNetCore.AutoHealthCheck.Tests/Infrastructure/Endpoints/EndpointMessageTranslatorTests.cs
