+++ tests/AspNetCore.AutoHealthCheck.Tests/Infrastructure/HealthCheckerTests.cs
