+++ tests/AspNetCore.AutoHealthCheck.Tests/Infrastructure/Middleware/AutoHealthCheckMiddlewareTests.cs
