+++ tests/AspNetCore.AutoHealthCheck.Tests/Infrastructure/ResultProcessor/HealtCheckResultProcessorNegativeTests.cs
