+++ tests/AspNetCore.AutoHealthCheck.Tests/Infrastructure/ResultProcessor/HealthCheckResultEndpointProcessorTests.cs
