+++ tests/AspNetCore.AutoHealthCheck.Tests/Infrastructure/ResultProcessor/HealthCheckResultProcessorTests.cs
